Separated views and updated urls
